creating new components + reorganizing folder and file structure + renaming files for context + use of assets folder instead of public folder
