+++ src/scss/_fonts.scss
@@ -1,18 +1,18 @@
 @font-face {
     font-family: 'Open Sans';
-    src: url('/fonts/OpenSans/OpenSans-Regular.ttf') format('truetype');
+    src: url('~@/assets/fonts/OpenSans/OpenSans-Regular.ttf') format('truetype');
     font-weight: normal;
     font-style: normal;
 }
 @font-face {
     font-family: 'Montserrat Medium';
-    src: url('/fonts/Montserrat/Montserrat-Medium.ttf') format('truetype');
+    src: url('~@/assets/fonts/Montserrat/Montserrat-Medium.ttf') format('truetype');
     font-weight: normal;
     font-style: normal;
 }
 @font-face {
     font-family: 'Montserrat SemiBold';
-    src: url('/fonts/Montserrat/Montserrat-SemiBold.ttf') format('truetype');
+    src: url('~@/assets/fonts/Montserrat/Montserrat-SemiBold.ttf') format('truetype');
     font-weight: normal;
     font-style: normal;
 }
